Move all files out of Product folder and delete folder
