+++ ui/src/Common/EditTagRow/EditTagRow.tsx
@@ -23,7 +23,7 @@ import {createDataTestId} from 'Utils/ReactUtils';
 import {RoleTagRequest, TagRequest} from 'Types/TagRequest';
 
 import {OptionType} from '../SelectWithHTMLOptions/SelectWithHTMLOptions';
-import ColorDropdown from 'Roles/ColorDropdown';
+import ColorDropdown from 'Common/EditTagRow/ColorDropdown/ColorDropdown';
 import {TagType} from 'SubHeader/SortingAndFiltering/FilterLibraries';
 import {Color} from 'Types/Color';
