+++ ui/src/PeopleMover/ArchiveProductsDrawer/ArchivedProduct/ArchivedProduct.tsx
@@ -18,7 +18,7 @@
 import React from 'react';
 import {useSetRecoilState} from 'recoil';
 import {ModalContentsState} from 'State/ModalContentsState';
-import ProductForm from 'Products/ProductForm/ProductForm';
+import ProductForm from 'Common/ProductForm/ProductForm';
 import {Product} from 'Types/Product';
 
 interface Props{
