+++ ui/src/PeopleMover/ProductList/ProductListGrouped/ProductListGrouped.tsx
@@ -25,7 +25,7 @@ import {ProductSortBy, ProductSortByState} from 'State/ProductSortByState';
 import {LocationsState} from 'State/LocationsState';
 import {ProductTagsState} from 'State/ProductTagsState';
 import {ModalContents} from 'State/ModalContentsState';
-import ProductForm from 'Products/ProductForm/ProductForm';
+import ProductForm from 'Common/ProductForm/ProductForm';
 import {UUIDForCurrentSpaceSelector} from 'State/CurrentSpaceState';
 import {Tag} from 'Types/Tag';
 import {Product} from 'Types/Product';
